login added, untested

+++ controllers/beerController.js
@@ -1,8 +1,18 @@
 const beerRouter = require('express').Router()
 const Beer = require('../models/beer')
+const User = require('../models/user')
 const axios = require('axios')
+const jwt = require('jsonwebtoken')
 
 // implement user and login check
+
+const getToken = (request) => {
+    const authorization = request.get('authorization')
+    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+        return authorization.substring(7)
+    }
+    return null
+}
 
 beerRouter.get('/', async (request, response) => {
     try {
@@ -35,16 +45,26 @@ beerRouter.post('/', async (request, response) => {
     const body = request.body
 
     try {
+        const token = getToken(request)
+        const decodedToken = jwt.verify(token, process.env.SEC)
+
+        if (!token || !decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+
         if (body === undefined) {
             return response.status(400).json({ error: 'content missing' })
         }
+
+        const user = await User.findById(decodedToken.id)
 
         const beer = new Beer({
             name: body.name,
             brewery: body.brewery,
             type: body.type,
             country: body.country,
-            alcohol_percent: body.alcohol_percent
+            alcohol_percent: body.alcohol_percent,
+            userWhoAdded: user
         })
 
         savedBeer = await beer.save()
@@ -55,9 +75,21 @@ beerRouter.post('/', async (request, response) => {
         response.status(500).json({ error: 'internal error' })
     }
 })
-
+// implement that only adder can delete. 
 beerRouter.delete('/:id', async (request, response) => {
     try {
+        const token = getToken(request)
+        const decodedToken = jwt.verify(token, process.env.SEC)
+
+        if (!token || !decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+
+        // experimental
+        if(beerToDelete.userWhoAdded.id !== decodedToken.id) {
+            return response.status(401).json({ error: 'only the creater can delete the beer' })
+        }
+
         const beerToDelete = await Beer.findByIdAndRemove(request.params.id)
         response.status(204).end()
 
@@ -71,6 +103,13 @@ beerRouter.put('/:id', async (request, response) => {
     const body = request.body
 
     try {
+        const token = getToken(request)
+        const decodedToken = jwt.verify(token, process.env.SEC)
+
+        if (!token || !decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+
         const beerToUpdate = {
             name: body.name,
             brewery: body.brewery,
